app/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.2.1
gunicorn==25.0.1
requests==2.32.3
